split up types modules
