+++ src/compiler/types/mod.rs
@@ -1,9 +1,19 @@
-mod lines_0_1000;
-pub use lines_0_1000::*;
-mod lines_1000_2000;
-pub use lines_1000_2000::*;
-mod lines_2000_4000;
-pub use lines_2000_4000::*;
+mod lines_0_500;
+pub use lines_0_500::*;
+mod lines_500_1000;
+pub use lines_500_1000::*;
+mod lines_1000_1500;
+pub use lines_1000_1500::*;
+mod lines_1500_2000;
+pub use lines_1500_2000::*;
+mod lines_2000_2500;
+pub use lines_2000_2500::*;
+mod lines_2500_3000;
+pub use lines_2500_3000::*;
+mod lines_3000_3500;
+pub use lines_3000_3500::*;
+mod lines_3500_4000;
+pub use lines_3500_4000::*;
 mod lines_4000_5000;
 pub use lines_4000_5000::*;
 mod lines_5000_5250;
@@ -12,7 +22,11 @@ mod lines_5250_5500;
 pub use lines_5250_5500::*;
 mod lines_5500_6000;
 pub use lines_5500_6000::*;
-mod lines_6000_8000;
-pub use lines_6000_8000::*;
+mod lines_6000_6500;
+pub use lines_6000_6500::*;
+mod lines_6500_7000;
+pub use lines_6500_7000::*;
+mod lines_7000_8000;
+pub use lines_7000_8000::*;
 mod lines_8000_end;
 pub use lines_8000_end::*;
